backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "bs4>=0.0.2",
    "cryptography>=43.0.3",
    "dj-rest-auth[with-social]==4.0.0",
    "django-allauth==0.52.0",
    "django-cors-headers==4.1.0",
    "django-debug-toolbar>=4.4.6",
    "django-livereload-server>=0.5.1",
    "django>=4.2.16",
    "djangorestframework-simplejwt==5.2.2",
    "djangorestframework==3.14.0",
    "huggingface-hub>=0.26.2",
    "loguru>=0.7.2",
    "psycopg2>=2.9.10",
    "selenium>=4.26.1",
    "setuptools>=75.3.0",
    "transformers>=4.46.3",
]
